검색 관련 컴포넌트 분리
- Search 컴포넌트 분리
- SearchItem 컴포넌트 분리
- PropTypes 검증
- Props 이름 수정
